trabjando con rutas dinamicas multiples
